backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
mysqlclient==2.2.0
python-dotenv==1.0.0
kafka-python==2.0.2
cryptography==41.0.3
pycryptodomex==3.19.0
python-magic==0.4.27
PyJWT==2.8.0
Werkzeug==2.3.7
